ajuste de erro quando não é informado nenhum item

+++ app/controllers/taxweb_widgets/widgets_controller.rb
@@ -26,14 +26,16 @@ module TaxwebWidgets
       widgets = params[:widgets]
       widget_name_user = []
       action_user = []
-      widgets.each do |widget|
-        widget_name, action = widget.split('|')
-        if widget_name && action
-          widget_name_user << widget_name
-          action_user << action
-          TaxwebWidgets::User.find_or_create_by(widget: widget_name, action: action, user_id: user_id)
+      if widgets.present? && user_id.present?
+        widgets.each do |widget|
+          widget_name, action = widget.split('|')
+          if widget_name && action
+            widget_name_user << widget_name
+            action_user << action
+            TaxwebWidgets::User.find_or_create_by(widget: widget_name, action: action, user_id: user_id)
+          end
+          TaxwebWidgets::User.where(user_id: current_user.id).where.not(widget: widget_name_user, action: action_user, ).destroy_all
         end
-        TaxwebWidgets::User.where(user_id: current_user.id).where.not(widget: widget_name_user, action: action_user, ).destroy_all
       end
       flash[:success] = 'Alterações foram salvas com sucesso!'
       redirect_to taxweb_widgets_path
